split up app/lib
