+++ app/src/test/kotlin/net/thauvin/erik/urlencoder/UrlEncoderTest.kt
@@ -17,14 +17,10 @@
 
 package net.thauvin.erik.urlencoder
 
-import net.thauvin.erik.urlencoder.UrlEncoder.decode
-import net.thauvin.erik.urlencoder.UrlEncoder.encode
 import net.thauvin.erik.urlencoder.UrlEncoder.processMain
 import net.thauvin.erik.urlencoder.UrlEncoder.usage
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertSame
 import org.junit.jupiter.api.Assertions.assertThrows
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
@@ -34,7 +30,6 @@ import org.junit.jupiter.params.provider.ValueSource
 import java.util.stream.Stream
 
 class UrlEncoderTest {
-    private val same = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVQXYZ0123456789-_."
 
     companion object {
         @JvmStatic
@@ -52,67 +47,6 @@ class UrlEncoderTest {
                 "\uD808\uDC00\uD809\uDD00\uD808\uDF00\uD808\uDD00", "%F0%92%80%80%F0%92%94%80%F0%92%8C%80%F0%92%84%80"
             )
         )
-    }
-
-    @ParameterizedTest(name = "decode({0}) should be {1}")
-    @MethodSource("validMap")
-    fun `Decode URL`(expected: String, source: String) {
-        assertEquals(expected, decode(source))
-    }
-
-    @ParameterizedTest(name = "decode({0})")
-    @MethodSource("invalid")
-    fun `Decode with Exception`(source: String) {
-        assertThrows(IllegalArgumentException::class.java, { decode(source) }, "decode($source)")
-    }
-
-    @Test
-    fun `Decode when None needed`() {
-        assertSame(same, decode(same))
-        assertEquals("", decode(""), "decode('')")
-        assertEquals(" ", decode(" "), "decode(' ')")
-    }
-
-    @Test
-    fun `Decode with Plus to Space`() {
-        assertEquals("foo bar", decode("foo+bar", true))
-        assertEquals("foo bar  foo", decode("foo+bar++foo", true))
-        assertEquals("foo  bar  foo", decode("foo+%20bar%20+foo", true))
-        assertEquals("foo + bar", decode("foo+%2B+bar", plusToSpace = true))
-        assertEquals("foo+bar", decode("foo%2Bbar", plusToSpace = true))
-    }
-
-    @ParameterizedTest(name = "encode({0}) should be {1}")
-    @MethodSource("validMap")
-    fun `Encode URL`(source: String, expected: String) {
-        assertEquals(expected, encode(source))
-    }
-
-    @Test
-    fun `Encode Empty or Blank`() {
-        assertTrue(encode("", allow = "").isEmpty(), "encode('','')")
-        assertEquals("", encode(""), "encode('')")
-        assertEquals("%20", encode(" "), "encode(' ')")
-    }
-
-    @Test
-    fun `Encode when None needed`() {
-        assertSame(same, encode(same))
-        assertSame(same, encode(same, allow = ""), "with empty allow")
-    }
-
-    @Test
-    fun `Encode with Allow Arg`() {
-        assertEquals("?test=a%20test", encode("?test=a test", allow = "=?"), "encode(x, =?)")
-        assertEquals("aaa", encode("aaa", "a"), "encode(aaa, a)")
-        assertEquals(" ", encode(" ",  " "), "encode(' ', ' ')")
-    }
-
-    @Test
-    fun `Encode with Space to Plus`() {
-        assertEquals("foo+bar", encode("foo bar", spaceToPlus = true))
-        assertEquals("foo+bar++foo", encode("foo bar  foo", spaceToPlus = true))
-        assertEquals("foo bar", encode("foo bar", " ", true))
     }
 
     @ParameterizedTest(name = "processMain(-d {1}) should be {0}")
