+++ lib/src/main/kotlin/net/thauvin/erik/urlencoder/UrlEncoderUtil.kt
@@ -19,7 +19,6 @@ package net.thauvin.erik.urlencoder
 
 import java.nio.charset.StandardCharsets
 import java.util.BitSet
-import kotlin.system.exitProcess
 
 /**
  * Most defensive approach to URL encoding and decoding.
@@ -37,13 +36,8 @@ import kotlin.system.exitProcess
  * @author Geert Bevin (gbevin(remove) at uwyn dot com)
  * @author Erik C. Thauvin (erik@thauvin.net)
  **/
-object UrlEncoder {
+object UrlEncoderUtil {
     private val hexDigits = "0123456789ABCDEF".toCharArray()
-    internal val usage =
-        "Usage : java -jar urlencoder-*all.jar [-ed] text" + System.lineSeparator() +
-                "Encode and decode URL components defensively." + System.lineSeparator() +
-                "  -e  encode (default) " + System.lineSeparator() +
-                "  -d  decode"
 
     // see https://www.rfc-editor.org/rfc/rfc3986#page-13
     // and https://url.spec.whatwg.org/#application-x-www-form-urlencoded-percent-encode-set
@@ -195,46 +189,5 @@ object UrlEncoder {
         }
 
         return out?.toString() ?: source
-    }
-
-    /**
-     * Encodes and decodes URLs from the command line.
-     *
-     * - `java -jar urlencoder-*all.jar [-ed] text`
-     */
-    @JvmStatic
-    fun main(args: Array<String>) {
-        try {
-            val result = processMain(args)
-            if (result.status == 1) {
-                System.err.println(result.output)
-            } else {
-                println(result.output)
-            }
-            exitProcess(result.status)
-        } catch (e: IllegalArgumentException) {
-            System.err.println("${UrlEncoder::class.java.simpleName}: ${e.message}")
-            exitProcess(1)
-        }
-    }
-
-    internal data class MainResult(var output: String = usage, var status: Int = 1)
-
-    internal fun processMain(args: Array<String>): MainResult {
-        val result = MainResult()
-        if (args.isNotEmpty() && args[0].isNotEmpty()) {
-            val hasDecode = (args[0] == "-d")
-            val hasOption = (hasDecode || args[0] == "-e")
-            if (hasOption && args.size == 2 || !hasOption && args.size == 1) {
-                val arg = if (hasOption) args[1] else args[0]
-                if (hasDecode) {
-                    result.output = decode(arg)
-                } else {
-                    result.output = encode(arg)
-                }
-                result.status = 0
-            }
-        }
-        return result
     }
 }
