+++ lib/src/test/kotlin/net/thauvin/erik/urlencoder/UrlEncoderUtilTest.kt
@@ -1,0 +1,114 @@
+/*
+ * Copyright 2001-2023 Geert Bevin (gbevin[remove] at uwyn dot com)
+ * Copyright 2022-2023 Erik C. Thauvin (erik@thauvin.net)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.thauvin.erik.urlencoder
+
+import net.thauvin.erik.urlencoder.UrlEncoderUtil.decode
+import net.thauvin.erik.urlencoder.UrlEncoderUtil.encode
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertSame
+import org.junit.jupiter.api.Assertions.assertThrows
+import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.Arguments.arguments
+import org.junit.jupiter.params.provider.MethodSource
+import java.util.stream.Stream
+
+class UrlEncoderUtilTest {
+    private val same = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVQXYZ0123456789-_."
+
+    companion object {
+        @JvmStatic
+        fun invalid() = arrayOf("sdkjfh%", "sdkjfh%6", "sdkjfh%xx", "sdfjfh%-1")
+
+        @JvmStatic
+        fun validMap(): Stream<Arguments> = Stream.of(
+            arguments("a test &", "a%20test%20%26"),
+            arguments(
+                "!abcdefghijklmnopqrstuvwxyz%%ABCDEFGHIJKLMNOPQRSTUVQXYZ0123456789-_.~=",
+                "%21abcdefghijklmnopqrstuvwxyz%25%25ABCDEFGHIJKLMNOPQRSTUVQXYZ0123456789-_.%7E%3D"
+            ),
+            arguments("%#okékÉȢ smile!😁", "%25%23ok%C3%A9k%C3%89%C8%A2%20smile%21%F0%9F%98%81"),
+            arguments(
+                "\uD808\uDC00\uD809\uDD00\uD808\uDF00\uD808\uDD00", "%F0%92%80%80%F0%92%94%80%F0%92%8C%80%F0%92%84%80"
+            )
+        )
+    }
+
+    @ParameterizedTest(name = "decode({0}) should be {1}")
+    @MethodSource("validMap")
+    fun `Decode URL`(expected: String, source: String) {
+        assertEquals(expected, decode(source))
+    }
+
+    @ParameterizedTest(name = "decode({0})")
+    @MethodSource("invalid")
+    fun `Decode with Exception`(source: String) {
+        assertThrows(IllegalArgumentException::class.java, { decode(source) }, "decode($source)")
+    }
+
+    @Test
+    fun `Decode when None needed`() {
+        assertSame(same, decode(same))
+        assertEquals("", decode(""), "decode('')")
+        assertEquals(" ", decode(" "), "decode(' ')")
+    }
+
+    @Test
+    fun `Decode with Plus to Space`() {
+        assertEquals("foo bar", decode("foo+bar", true))
+        assertEquals("foo bar  foo", decode("foo+bar++foo", true))
+        assertEquals("foo  bar  foo", decode("foo+%20bar%20+foo", true))
+        assertEquals("foo + bar", decode("foo+%2B+bar", plusToSpace = true))
+        assertEquals("foo+bar", decode("foo%2Bbar", plusToSpace = true))
+    }
+
+    @ParameterizedTest(name = "encode({0}) should be {1}")
+    @MethodSource("validMap")
+    fun `Encode URL`(source: String, expected: String) {
+        assertEquals(expected, encode(source))
+    }
+
+    @Test
+    fun `Encode Empty or Blank`() {
+        assertTrue(encode("", allow = "").isEmpty(), "encode('','')")
+        assertEquals("", encode(""), "encode('')")
+        assertEquals("%20", encode(" "), "encode(' ')")
+    }
+
+    @Test
+    fun `Encode when None needed`() {
+        assertSame(same, encode(same))
+        assertSame(same, encode(same, allow = ""), "with empty allow")
+    }
+
+    @Test
+    fun `Encode with Allow Arg`() {
+        assertEquals("?test=a%20test", encode("?test=a test", allow = "=?"), "encode(x, =?)")
+        assertEquals("aaa", encode("aaa", "a"), "encode(aaa, a)")
+        assertEquals(" ", encode(" ",  " "), "encode(' ', ' ')")
+    }
+
+    @Test
+    fun `Encode with Space to Plus`() {
+        assertEquals("foo+bar", encode("foo bar", spaceToPlus = true))
+        assertEquals("foo+bar++foo", encode("foo bar  foo", spaceToPlus = true))
+        assertEquals("foo bar", encode("foo bar", " ", true))
+    }
+}
