+++ settings.gradle.kts
@@ -20,6 +20,9 @@ dependencyResolutionManagement {
   }
 }
 
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+
 include(
+  ":app",
   ":lib",
 )
